Backfill daily candles for the whole market
- backfill_candles.py now fetches the full ticker list (or --symbols),
  with --days and --limit options instead of a hardcoded symbol/range
- Insert with ON CONFLICT DO NOTHING in one statement per symbol instead
  of a SELECT per day; commit per symbol so progress survives a crash
- Retry Fugle calls with exponential backoff on HTTP 429; failed symbols
  are logged and skipped
- stock_candles.volume -> BigInteger (some ETFs exceed int32 daily volume)

+++ app/models/stock_candle.py
@@ -1,6 +1,6 @@
 from datetime import date as date_type
 
-from sqlalchemy import Date, Integer, Numeric, String, UniqueConstraint
+from sqlalchemy import BigInteger, Date, Integer, Numeric, String, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.core.database import Base
@@ -17,6 +17,6 @@ class StockCandle(Base):
     high: Mapped[float] = mapped_column(Numeric(12, 4), nullable=False)
     low: Mapped[float] = mapped_column(Numeric(12, 4), nullable=False)
     close: Mapped[float] = mapped_column(Numeric(12, 4), nullable=False)
-    volume: Mapped[int] = mapped_column(Integer, nullable=False)
+    volume: Mapped[int] = mapped_column(BigInteger, nullable=False)
     turnover: Mapped[float] = mapped_column(Numeric(16, 2), nullable=False)
     change: Mapped[float] = mapped_column(Numeric(12, 4), nullable=False)
